prototyping

+++ include/EvoAPI.hpp
@@ -17,8 +17,8 @@ class EvoAPI {
     static std::shared_ptr<spdlog::logger> logger;
 
     // algorithm boundary conditions
-    int generation_size_limit, generation_count_limit, interaction_cols, mutation_rate, island_count, migration_ratio, migration_interval;
-    int global_generation_size_limit, migrants_count;
+    size_t generation_size_limit, generation_count_limit, interaction_cols, mutation_rate, island_count, migration_ratio, migration_interval;
+    size_t global_generation_size_limit, migrants_count;
 
     // preprocessed input data
     Eigen::MatrixXd x, y;
@@ -35,7 +35,8 @@ class EvoAPI {
     void create_regression_input(std::tuple<int, std::vector<double>>);
     Transform::EvoDataSet get_dataset();
 
-    // concurrent random engines
+    // concurrent random engines 
+    // TODO move to EvoLibrary
     std::vector<XoshiroCpp::Xoshiro256Plus> create_random_engines(int count);
 
     // fitness & generation postprocessing
